backend/mlModels/Road construction/analysis/requirements.txt
ultralytics
pillow
